allow use of external fencing devices
A node now can be fenced with the use of external hardware fence
devices.
Those device can be configured at /etc/pve/ha/fence.cfg
also the fencing option in the datacenter configuration file must
be set to either 'hardware' or 'both', else configured devices
will *not* be used.

If hardware is selected as mode a valid device config *must* be
made, the fencing will not be marked as successful even if the CRM
could theoretical acquire the lock from the failed node!
This is done as some setups may require a HW fence agent to cut the
node off and where a watchdog which resets the node may be
dangerous.

We always *must* acquire the log before we can mark the failed node
as fenced and place it into the 'unknown' state and recover its
services.

The CRM bails out in case of an lost manager lock event where
$manager->cleanup() gets called.
There we kill all remaining open fence processes, if any,
and reset the fence status.

The currents masters manager class processes the running fencing
jobs, this means picking up finished fence workers and evaluating
their result.

Now regressions test with faked virtual HW fence devices are also
possible.
The current virtual devices succeed always, this will be changed
in a future patch to allow testing of more (dangerous) corner cases.

Device can be configured in the testdir/fence.cfg file and follow
the exactly same format as the real ones (see man dlm.conf)