always fence nodes on dead LRM
fixes a recovery failure if a node starts up with a dead/broken LRM
but working corosync.

So while its quorate it doesn't do anything but the CRM won't fence
it as our "last_online" timestamp only checks if quorate, not if the
HA manager is actually working.

Can be reproduced with having a active node with services, simply
disable the lrm:
$ systemctl disable pve-ha-lrm
and then reboot.
(this would simulate a broken update/reboot)
So the node gets up again and gains quorum but the LRM does not
start and thus no service gets started/migrated/... fencing is
appropriate for such a situation.