test_fence_config: add deep object compare method
Can compare Objects with arbitary nested data structures of the
types hash, array, and scalar.
It makes an depth first search with some early abortion checks on
the roots of subtrees (is the length and type the same).

Will be used to compare configs for the parser self check test in
the next patch.