fix LRM error on corrupted manager status
If the manager stateus file gets corrupted, meaning no valid JSON or
emtpy both LRM and CRM may crash here as decode_json throws an error
then. This then triggers the watchdog if already opened, as it won't
get reset anymore.

Reproducable with:
$ > /etc/pve/ha/manager_status

Put an eval around the decode json call to catch all possible errors
and if such an error happens recover gracefully, i.e. just return an
empty JSON object. As the current master has the state saved in
memory he writes it back on the next loop iteration.
If no CRM is manager at the time it is able to rebuild the state
from scratch.