add check if a service is relocatable and assert it on recovery
Add a basic check if a service is relocatable, i.e. is bound to
local resources from a node.

Assert this check when a service gets recovered after a failed node
to avoid moving the service only so that it fails to start, even if
we knew beforehand that it cannot start.

Place such services in the error state as they need manual
intervention.