Fix #1160: add template resource
This is a very limited resource representing both LXC and QEMU
templates.
Methods like start, stop and check_running do nothing or tell the
LRM want he wants to hear.

While templates are a questionable HA resource they are already used
as such (see bug 1160) and prohibiting that usage is a draw back for
those users and has no real value for us.